requirements.txt
aws-cdk.core==1.169.0
aws-cdk.aws_glue==1.169.0
aws-cdk.pipelines==1.169.0
aws_cdk.aws_s3_deployment==1.169.0
pyyaml==5.1

